First working draft
First version that works with varying sized browser windows and properly respects rules of tile movement.

Object of the lesson: learn to use jQuery to respond to clicks and move the tiles around.

The tile themselves are very basic. A possible future project would be to play with CSS and make it prettier.